requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1

# Authentication
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# AI/ML
google-generativeai==0.3.1
openai==1.3.7  # For Whisper API (optional)

# Media Processing
pillow==10.1.0
python-magic==0.4.27
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# Testing (dev)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Type hints
typing-extensions==4.9.0

